TASK: Separate Booting\Scripts and CLI subrequest calls
This introduces `Neos\Flow\Core\PhpCliCommandHandler` to run
CLI subrequests instead of using `\Neos\Flow\Core\Booting\Scripts` as
that has grown way to big over the years and this single responsibility
can be cleanly separated into a different class.

Some minimal cleanup was done along the way.

The old `@api` methods still exist in Scripts as deprecated shims
so this should not be breaking for userland code.
